requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-filter==23.3
drf_yasg==1.21.7
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-redis==5.4.0
django-tinymce==3.6.1